Adds angle and response onto the KeyPoint generated by the MSER
Angle has been extracted from ellipse angle, which is obtained from the MSER and fitEllipse function of OpenCV.
Response is initially evaluated has the máximum pixel intensity, further evaluating criteria need to be developed.

TODO:
Check KeyPoint angle and vertical scale applied to the detected region.
Obtain scale of the given MSER in order to include it onto the KeyPoint gerenated for the SIFT descriptor.